+++ CHANGELOG.md
@@ -88,6 +88,7 @@ previous versions. You will need to upgrade your old database manually or using
 
 ### Plotting
 
+- Show the station radiated energy (Er) value on the station spectra plots
 - Station maps: exclude outliers when computing colorbar limits
 - Station maps: improved computation of bounding box for regional and
   teleseismic events